requirements.txt
fastapi==0.115.14
pymongo==4.13.2
uvicorn==0.34.3
python-dotenv
firebase-admin==6.9.0
pyjwt
mangum